_materiali/progetti corretti/ft_trascenders2/study/requirements.txt
django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
django-cors-headers==4.3.1
Pillow==10.2.0
requests==2.31.0
PyJWT
python-dotenv
daphne==3.0.1
channels_redis==4.2
channels==4.0.0